Move Progress struct to pkg/cmd/core
